vim: start reworking language-specific plugin configuration
Source vimrc.lua in Neovim, where built-in NeoVim LSP server is configured.

Do not use LanguageClient for C++ for now, but do use vim-lsp and
built-in NeoVim LSP client, as well as vim-ccls.

clangd and ccls will be LSP servers for C++ sources.

Add ALE configuration stub for C++ but keep the ale plugin disabled for now.